Moved node operator out off kafka dir
