+++ cmd/insprd/operators/nodes/converter.go
