+++ cmd/insprd/operators/nodes/converter_test.go
