+++ cmd/insprd/operators/nodes/interfaces.go
