+++ cmd/insprd/operators/nodes/k8s_variables.go
