+++ cmd/insprd/operators/nodes/node_operator.go
