+++ cmd/insprd/operators/nodes/node_operator_test.go
