+++ examples/k8s_op_demo/main.go
@@ -5,7 +5,7 @@ import (
 	"log"
 
 	"github.com/inspr/inspr/cmd/insprd/memory/tree"
-	"github.com/inspr/inspr/cmd/insprd/operators/kafka/nodes"
+	"github.com/inspr/inspr/cmd/insprd/operators/nodes"
 	"github.com/inspr/inspr/pkg/meta"
 	"github.com/inspr/inspr/pkg/utils"
 )
